requirements.txt
python
azure-storage-blob
python-dotenv
